pyproject.toml
[tool.poetry]
name = "arbitrage-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
taskiq = {extras = ["reload"], version = "^0.8.8"}
taskiq-redis = "^0.4.0"
taskiq-aio-pika = "^0.4.0"
taskiq-pipelines = "^0.1.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.20"}
asyncio = "^3.4.3"
psycopg2-binary = "^2.9.7"
asyncpg = "^0.28.0"
httpx = "^0.25.0"
tldextract = "^3.5.0"
taskiq-dependencies = "^1.4.0"
simplejson = "^3.19.1"
python-dateutil = "^2.8.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-env = "^1.0.1"
respx = "^0.20.2"
pytest-cov = "^4.1.0"
pytest-dependency = "^0.5.1"

[tool.black]
line-length = 88
target-version = ['py311']


[tool.usort]

[tool.usort.known]
first_party=["src", "config", "tests"]


[tool.pytest.ini_options]
env = [
    "ENVIRONMENT=pytest",
]

[tool.coverage.run]
omit = [
    # omit tests
    "tests/*",
    # broker
    "src/worker/broker.py",
    # omit config
    "config.py",
    "conftest.py",
]

[tool.coverage.report]
skip_empty = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
